requirements.txt
streamlit
crewai
crewai-tools
python-dotenv
langchain
langchain-openai
langchain-community
chromadb
requests